Validate content_id when importing content
This allows to return a meaningful error if the content provided is missing a
content_id.

+++ app/commands/v2/import.rb
@@ -3,6 +3,17 @@ module Commands
     class Import < BaseCommand
 
       def call
+        unless UuidValidator.valid?(payload[:content_id])
+          raise CommandError.new(
+          code: 422,
+          error_details: {
+            error: {
+              code: 422,
+              message: "Content id not valid",
+              fields: "content_id",
+            }
+          })
+        end
         ContentItem.transaction do
           delete_all(payload[:content_id])
           all_content_items.map.with_index do |event, index|
